fix: use Decimal for design_hours, remove redundant selectinload

+++ backend/app/schemas/design_details.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from pydantic import BaseModel
 
 
@@ -6,7 +8,7 @@ class DesignDetailsCreate(BaseModel):
 
 
 class DesignDetailsResponse(BaseModel):
-    design_hours: float | None = None
+    design_hours: Decimal | None = None
     design_version_count: int = 0
     revision_count: int = 0
     proofing_data: dict | None = None
